# v3/api/openapi.yaml
openapi: 3.1.0
info:
  title: Pymes API
  version: 1.0.0-draft
  description: API pública de Pymes v3. Los importes se representan como strings decimales.
security:
  - clerkBearer: []
  - clerkSessionCookie: []
paths:
...
  /api/v1/organizations/{organizationId}/calendars/google/oauth/start:
    post:
      operationId: startGoogleCalendarOAuth
      summary: Inicia una conexión OAuth de Google Calendar para la organización.
      parameters:
        - { $ref: '#/components/parameters/OrganizationId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GoogleCalendarOAuthStartInput' }
      responses:
        '201':
          description: Estado OAuth durable, de propósito único y ligado a la sesión Clerk.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GoogleCalendarOAuthStart' }
        '400': { $ref: '#/components/responses/InvalidCommand' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/DomainError' }
        '503': { $ref: '#/components/responses/AuthUnavailable' }
  /api/v1/calendars/google/oauth/callback:
    get:
      operationId: completeGoogleCalendarOAuth
      summary: Completa OAuth directamente en el BFF sin entregar el código a la web.
      description: >-
        Google redirige a esta ruta global. El BFF vuelve a autenticar la misma
        sesión Clerk y consume el state una sola vez, ligado a tenant, actor,
        sesión y vencimiento.
      parameters:
        - name: state
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 1024 }
        - name: code
          in: query
          required: false
          schema: { type: string, minLength: 1, maxLength: 4096 }
        - name: error
          in: query
          required: false
          schema: { type: string, minLength: 1, maxLength: 255 }
      responses:
        '200':
          description: Conexión creada o reconciliada.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CalendarConnection' }
        '400': { $ref: '#/components/responses/InvalidCommand' }
        '401':
          description: La conexión requiere una nueva autorización.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '410':
          description: El estado OAuth venció.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/DomainError' }
        '503':
          description: Google Calendar o IAM no está disponible.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/organizations/{organizationId}/calendars/connections:
    get:
      operationId: listCalendarConnections
      summary: Lista conexiones Calendar sin exponer grants ni tokens.
      parameters:
        - { $ref: '#/components/parameters/OrganizationId' }
      responses:
        '200':
          description: Conexiones visibles dentro de la organización.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CalendarConnection' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/AuthUnavailable' }
  /api/v1/organizations/{organizationId}/calendars/connections/{connectionId}:
    delete:
      operationId: disconnectCalendarConnection
      summary: Revoca y desconecta una conexión Calendar de la organización.
      parameters:
        - { $ref: '#/components/parameters/OrganizationId' }
        - name: connectionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Conexión revocada o desconexión ya convergida. }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503':
          description: Google Calendar o IAM no está disponible.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    clerkBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT de sesión Clerk validado por el BFF.
    clerkSessionCookie:
      type: apiKey
      in: cookie
      name: __session
      description: >-
        JWT de sesión Clerk aceptado sólo cuando navegador y BFF comparten el
        mismo origen; permite completar redirects OAuth sin exponer el código.
  parameters:
    OrganizationId: { name: organizationId, in: path, required: true, schema: { type: string } }
...
  responses:
    InvalidCommand:
      description: Comando inválido o cabeceras idempotentes ausentes.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Forbidden:
      description: La sesión Clerk no pertenece a la organización o el rol no autoriza la operación.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    DomainError:
      description: La organización no está lista o una regla comercial impide el comando.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: El recurso no existe en la organización autenticada.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    AuthUnavailable:
      description: Clerk/IAM no está configurado en el BFF.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [code]
      properties:
        code:
          type: string
          enum:
            - AUTH_NOT_CONFIGURED
            - FORBIDDEN
            - ORG_NOT_PROVISIONED
            - NOT_FOUND
            - IDEMPOTENCY_KEY_REQUIRED
            - SOURCE_VERSION_REQUIRED
            - IDEMPOTENCY_KEY_REUSED
            - DOCUMENT_NOT_REVERSIBLE
            - INVALID_APPLICATION_DOCUMENT
            - INVALID_SOURCE_DOCUMENT
            - OPEN_ITEM_AMOUNT_EXCEEDED
            - PERIOD_LOCKED
            - ACCOUNTING_ADJUSTMENT_NOT_ALLOWED
            - CREDENTIAL_NOT_FOUND
            - CREDENTIAL_NOT_READY
            - CREDENTIAL_VERSION_CONFLICT
            - CERTIFICATE_INVALID
            - CERTIFICATE_CUIT_MISMATCH
            - CERTIFICATE_ENVIRONMENT_MISMATCH
            - POINT_OF_SALE_NOT_VALIDATED
            - AUTHORITY_TIMEOUT
            - FISCAL_UNAVAILABLE
            - VALIDATION_ERROR
            - COMMAND_REJECTED
            - NOTIFICATIONS_UNAVAILABLE
            - PERGO_WEBHOOK_NOT_CONFIGURED
            - PERGO_WEBHOOK_SIGNATURE_INVALID
            - PERGO_WEBHOOK_INVALID
            - NOTIFICATION_NOT_FOUND
            - OAUTH_PROVIDER_DENIED
            - OAUTH_STATE_INVALID
            - OAUTH_STATE_EXPIRED
            - FEATURE_DISABLED
            - FEATURE_VERSION_CONFLICT
            - CALENDAR_REAUTH_REQUIRED
            - CALENDAR_PROVIDER_UNAVAILABLE
            - INTERNAL_ERROR
...
    GoogleCalendarOAuthStartInput:
      type: object
      additionalProperties: false
      required: [time_zone, free_busy_enabled, meet_enabled]
      properties:
        time_zone:
          type: string
          minLength: 1
          maxLength: 255
          description: Zona IANA usada por el calendario secundario.
        free_busy_enabled: { type: boolean }
        meet_enabled: { type: boolean }
    GoogleCalendarOAuthStart:
      type: object
      additionalProperties: false
      required: [connection_id, authorization_url, expires_at]
      properties:
        connection_id: { type: string, format: uuid }
        authorization_url: { type: string, format: uri }
        expires_at: { type: string, format: date-time }
    CalendarConnection:
      type: object
      additionalProperties: false
      required:
        - id
        - provider
        - status
        - calendar_connected
        - time_zone
        - free_busy_enabled
        - meet_enabled
        - version
      properties:
        id: { type: string, format: uuid }
        provider: { type: string, const: google }
        status:
          type: string
          enum: [pending, active, reauth_required, revoked]
        calendar_connected: { type: boolean }
        time_zone: { type: string, minLength: 1, maxLength: 255 }
        free_busy_enabled: { type: boolean }
        meet_enabled: { type: boolean }
        access_token_expires_at: { type: string, format: date-time }
        version: { type: integer, minimum: 1 }
...